% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Cleans out unseen triangles}
\usage{
clean(df, lower_x, lower_y, upper_x, upper_y)
}
\arguments{
\item{df}{A dataset of triangles}

\item{lower_x}{x position of the bottom left corner}

\item{lower_y}{y position of the bottom left corner}

\item{upper_x}{x position of the top right corner}

\item{upper_y}{y position of the top right corner}
}
\value{
A data set of filtered triangles
}
\description{
This function removes triangles that won't appear in a specified rectangle.
The purpose is to reduce the number of triangles to subdivide if they won't be
seen in the final output. If a triangle is fully above, below, or to the side
of the rectangle defined by input parameters, then it is removed from the data set.
The parameters \code{lower_x} with \code{lower_y} are the bottom
left corner and \code{upper_x} with \code{upper_y} are the top right corner.
Please not this might not remove all triangles that won't be seen as some can
extend across the diagonals and be missed by this filtering.
}
\examples{
return_starting_circle(0, 0, 0, 5, "t") \%>\% clean(-3, -3, -.5, -.5)
}
