% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_starting_circle.R
\name{return_starting_circle}
\alias{return_starting_circle}
\title{Get a starting circle of rhombi}
\usage{
return_starting_circle(point_x, point_y, rotation, side_length, shape)
}
\arguments{
\item{point_x}{Starting x position for first point}

\item{point_y}{Starting y position for first point}

\item{rotation}{Rotation about the starting position in degrees}

\item{side_length}{Length of one outside edge}

\item{shape}{Ether "t" for thin rhombus or "T" for thick rhombus}
}
\value{
Data set of triangles that join to create a circle of rhombi
}
\description{
This function returns a data set of ten observations. Each observation is a
triangle for one half of a rhombus. The location, direction, and shape of the
rhombus is determined through the parameters.
}
\details{
The center point of the circle is given by the \code{point_x} and
\code{point_y} parameters. This will appear in the data set as
\code{point1_x} and \code{point1_y}. The \code{rotation} in degrees will turn
the rhombi around this point while \code{side_length} will change the size.
Finally, \code{shape} will provide either a thick or thin rhombus. A "t" for
the shape parameter will yield a thin rhombus. Anything else will provide a
thick one, but traditionally a "T" is used. Thin triangles all point into the
circle (ten rhombi halfs) while thick triangles alternate pointing to each
other (five full rhombi).
}
\examples{
return_starting_circle(0, 0, 0, 1, "T")
return_starting_circle(5, 5, 180, 3, "t")
}
