% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdivide.R
\name{subdivide}
\alias{subdivide}
\title{Subdivides one triangle}
\usage{
subdivide(shape, point1_x, point1_y, point2_x, point2_y, point3_x, point3_y)
}
\arguments{
\item{shape}{Either "t" or "T" for a thin or thick triangle}

\item{point1_x}{x position of the first point}

\item{point1_y}{y position of the first point}

\item{point2_x}{x position of the middle point}

\item{point2_y}{y position of the middle point}

\item{point3_x}{x position of the last point}

\item{point3_y}{y position of the last point}
}
\value{
A data set of triangles
}
\description{
This function subdivides one triangle into either two or three more based on
the Penrose tiling rules.
}
\details{
If the shape is "t", there will be two triangles. If the shape is "T" or
anything else, there will be three triangles. The \code{point2_x} and
\code{point2_y} parameters determine the middle of the triangle. Which side
\code{point1_x} and \code{point1_y} versus \code{point3_x} and
\code{point3_y} determines if it is a left or right triangle. This function
is mostly used internally.
}
