% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitution.R
\name{substitution}
\alias{substitution}
\title{Apply substitution rules to a data set of rhombi/triangles}
\usage{
substitution(df)
}
\arguments{
\item{df}{A data set of triangles}
}
\value{
A data set of subdivided triangles
}
\description{
This function replaces loops through a data set of triangles, subdivides
them, then updates the triangle ids and rhombus ids.
}
\details{
The data set has to have \code{shape}, \code{point1_x}, \code{point1_y},
\code{point2_x}, \code{point2_y}, \code{point3_x}, and \code{point3_y}. These
columns will be updated. In addition to those, any column titles
\code{triangle} or \code{rhombus} will be updated to serve as identification
columns.

There is a section of this code that will snap the points to a grid in an
attempt to fix numerical stability issues. There is potential for this to
cause errors but almost always fixes more problems than it creates.
}
\examples{
return_rhombus(0, 0, -90 + 36, 1, "T") \%>\% substitution()

}
